+++ frontend/src/components/faculty/RejectedVerification.tsx
@@ -4,23 +4,11 @@ import { useEffect, useState, useMemo, useRef } from "react";
 import Pagination from "./Pagination";
 import { CheckCircle, Loader2 } from "lucide-react";
 import MannualAlert from "./MannualAlert";
+import { Student ,Subject , CertificateDetails, CertificateApiResponse, RejectedRequestWithDetails } from "../../types/faculty/MannualVerification";
 
-export type Subject = {
-  id: string;
-  name: string;
-  subject_code: string;
-  teacher_id: string;
-};
 
 export type ApiResponse = {
   subjects: Subject[];
-};
-
-export type Student = {
-  id: string;
-  name: string;
-  email: string;
-  roll_number: string;
 };
 
 export type Request = {
@@ -35,43 +23,6 @@ export type Request = {
 
 export type ApiResponseCSV = {
   requests: Request[];
-};
-
-export type CertificateDetails = {
-  uploaded_certificate: {
-    student_name: string;
-    roll_no: string;
-    marks: string;
-    course_name: string;
-    course_period: string;
-    file_url: string;
-  };
-  verification_certificate: {
-    student_name: string;
-    roll_no: string;
-    marks: string;
-    course_name: string;
-    course_period: string;
-    file_url: string;
-  };
-  subject_name: string;
-  remark: string;
-};
-
-export type CertificateApiResponse = {
-  message: string;
-  data: CertificateDetails;
-};
-
-export type RejectedRequestWithDetails = {
-  id: string;
-  student: Student;
-  subject: Subject;
-  status: "rejected";
-  verified_total_marks: string;
-  created_at: string;
-  due_date: string;
-  certificate_details: CertificateDetails | null;
 };
 
 const headings = [
